+++ test/2_fund_user_exchange_gas.js
@@ -119,7 +119,7 @@ contract("RariFundProxy", accounts => {
   it("should withdraw and exchange all input currencies without using too much gas", async () => {
     let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
     let fundTokenInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundToken.at(process.env.UPGRADE_FUND_TOKEN_ADDRESS) : RariFundToken.deployed());
-    let fundProxyInstance = await RariFundProxy.deployed();
+    let fundProxyInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundProxy.at(process.env.UPGRADE_FUND_PROXY_ADDRESS) : RariFundProxy.deployed());
 
     // Exchange data
     var inputCurrencyCodes = [];
@@ -174,9 +174,6 @@ contract("RariFundProxy", accounts => {
       protocolFeeBNs.push(web3.utils.toBN(protocolFee));
       totalProtocolFeeBN.iadd(web3.utils.toBN(protocolFee));
     }
-
-    // Approve RFT to RariFundManager
-    await fundTokenInstance.approve(RariFundManager.address, web3.utils.toBN(2).pow(web3.utils.toBN(256)).subn(1));
     
     // Fill 0x orders
     var result = await fundProxyInstance.withdrawAndExchange(inputCurrencyCodes, inputAmountBNs, "0x0000000000000000000000000000000000000000", allOrders, allSignatures, makerAssetFillAmountBNs, protocolFeeBNs, { value: totalProtocolFeeBN, gas: 8e6 });