+++ test/6_fund_fees.js
@@ -16,7 +16,6 @@ const RariFundManager = artifacts.require("RariFundManager");
 const RariFundToken = artifacts.require("RariFundToken");
 
 if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) {
-  RariFundController.address = process.env.UPGRADE_FUND_CONTROLLER_ADDRESS;
   RariFundManager.address = process.env.UPGRADE_FUND_MANAGER_ADDRESS;
   RariFundToken.address = process.env.UPGRADE_FUND_TOKEN_ADDRESS;
 }
@@ -24,7 +23,7 @@ if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) {
 // These tests expect the owner and the fund rebalancer of RariFundController and RariFundManager to be set to process.env.DEVELOPMENT_ADDRESS
 contract("RariFundManager", accounts => {
   it("should deposit to pools, set the interest fee rate, wait for interest, set the master beneficiary of interest fees, and deposit fees", async () => {
-    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundControllerInstance = await RariFundController.deployed();
     let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
     let fundTokenInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundToken.at(process.env.UPGRADE_FUND_TOKEN_ADDRESS) : RariFundToken.deployed());